bot2/requirements.txt
# Core ML/Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0

# Data Sources
polygon-api-client>=1.12.0
newsapi-python>=0.2.7
praw>=7.7.0
fredapi>=0.5.1
pytrends>=4.9.2

# Technical Analysis
ta-lib>=0.4.0  # May require manual installation on Windows
TA-Lib>=0.4.28  # Alternative pure Python implementation

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3
tqdm>=4.66.0
requests>=2.31.0

# Development/Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional (for advanced features)
yfinance>=0.2.28
beautifulsoup4>=4.12.0

